requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.4.0
Django==4.2.7
django-browser-reload==1.12.0
django-tailwind==3.6.0
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.1.0
